refactoring meta to app-template and data instead of scaffolding and added tests
