+++ packages/cli/src/data/queries/config.gql
@@ -9,7 +9,8 @@ query {
       rel,
       content,
       property,
-      value
+      value,
+      href
     },
     publicPath,
     title,
