+++ packages/cli/test/cases/build.config.theme/src/templates/page-template.js
@@ -1,13 +1,10 @@
 import { html, LitElement } from 'lit-element';
 import '../styles/my-brand.css';
 MDIMPORT;
-METAIMPORT;
-METADATA;
 
 class PageTemplate extends LitElement {
   render() {
     return html`
-      METAELEMENT
       <div class='wrapper'>
         <div class='page-template content owen-test'>
           <entry></entry>
