+++ packages/cli/test/cases/build.default.workspace-template-page-style/src/templates/page-template.js
@@ -2,8 +2,6 @@ import { html, LitElement } from 'lit-element';
 import '../styles/theme.css';
 import css from '../styles/style.css';
 MDIMPORT;
-METAIMPORT;
-METADATA;
 
 class PageTemplate extends LitElement {
   render() {
@@ -11,7 +9,6 @@ class PageTemplate extends LitElement {
       <style>
         ${css}
       </style>
-      METAELEMENT
       <div class='wrapper'>
         <div class='page-template content owen-test'>
           <entry></entry>
