+++ packages/cli/test/cases/build.default.workspace-template-page/src/templates/page-template.js
@@ -1,12 +1,9 @@
 import { html, LitElement } from 'lit-element';
 MDIMPORT;
-METAIMPORT;
-METADATA;
 
 class PageTemplate extends LitElement {
   render() {
     return html`
-      METAELEMENT
       <div class='wrapper'>
         <div class='page-template blog-content content owen-test'>
           <entry></entry>
