+++ packages/cli/test/cases/build.default.workspace-user-directory-mapping/src/templates/page-template.js
@@ -1,14 +1,10 @@
 import { html, LitElement } from 'lit-element';
 import '../components/header/header';
-
 MDIMPORT;
-METAIMPORT;
-METADATA;
 
 class PageTemplate extends LitElement {
   render() {
     return html`
-      METAELEMENT
       <div class='wrapper'>
         <div class='page-template blog-content content owen-test'>
           <entry></entry>
