+++ www/pages/getting-started/key-concepts.md
@@ -42,13 +42,10 @@ For reference, here's what the default page template included by Greenwood looks
 ```render javascript
 import { html, LitElement } from 'lit-element';
 MDIMPORT;
-METAIMPORT;
-METADATA;
 
 class PageTemplate extends LitElement {
   render() {
     return html\`
-      METAELEMENT
       <div class='wrapper'>
         <div class='page-template content'>
           <entry></entry>
@@ -61,7 +58,7 @@ class PageTemplate extends LitElement {
 customElements.define('page-template', PageTemplate);
 ```
 
-> Don't worry too much about the capitalized expressions, this is discussed in more detail in our [docs](/docs/template/).  Also, as seen here, Greenwood provides a version of [**LitElement**](https://lit-element.polymer-project.org/) by default that you can use for your own components if you would like.
+> Don't worry too much about the capitalized expression, this is discussed in more detail in our [docs](/docs/template/).  Also, as seen here, Greenwood provides a version of [**LitElement**](https://lit-element.polymer-project.org/) by default that you can use for your own components if you would like.
 
 
 ### Pages
